// src/domain/interfaces/balance.interface.ts
export interface IBalance {
  CODIGO_SALDO: number;
  ANO: number;
  FINANCIADO_ID: number;
  CODIGO_FINANCIADO: string;
  DESCRIPCION_FINANCIADO: string;
  CODIGO_ICP: string;
  CODIGO_SECTOR: string;
  CODIGO_PROGRAMA: string;
  CODIGO_SUBPROGRAMA: string;
  CODIGO_PROYECTO: string;
  CODIGO_ACTIVIDAD: string;
  CODIGO_OFICINA: string;
  CODIGO_ICP_CONCAT: string;
  DENOMINACION_ICP: string;
  UNIDAD_EJECUTORA: string;
  CODIGO_PUC: string;
  CODIGO_GRUPO: string;
  CODIGO_PARTIDA: string;
  CODIGO_GENERICA: string;
  CODIGO_ESPECIFICA: string;
  CODIGO_SUBESPECIFICA: string;
  CODIGO_NIVEL5: string;
  CODIGO_PUC_CONCAT: string;
  DENOMINACION_PUC: string;
  PRESUPUESTADO: number;
  ASIGNACION: number;
  BLOQUEADO: number;
  MODIFICADO: number;
  AJUSTADO: number;
  VIGENTE: number;
  COMPROMETIDO: number;
  POR_COMPROMETIDO: number;
  DISPONIBLE: number;
  CAUSADO: number;
  POR_CAUSADO: number;
  PAGADO: number;
  POR_PAGADO: number;
  CODIGO_EMPRESA: number;
  CODIGO_PRESUPUESTO: number;
  FECHA_SOLICITUD: Date;
  DESCRIPTIVA_FINANCIADO: string;
  SEARCH_TEXT: string;
}